requirements.txt
flask==2.3.3
groq==0.4.1
transformers>=4.35.0
torch>=2.0.0
opencv-python==4.8.1.78
opencv-contrib-python
pillow==10.0.1
matplotlib==3.7.2
numpy==1.24.3
werkzeug==2.3.7
deep-sort-realtime==1.3.2
python-dotenv==1.0.0 